+++ canvasapi/__init__.py
@@ -4,4 +4,4 @@ from canvasapi.canvas import Canvas
 
 __all__ = ["Canvas"]
 
-__version__ = "2.0.0"
+__version__ = "2.1.0"