+++ canvasapi/pairing_code.py
@@ -1,0 +1,6 @@
+from canvasapi.canvas_object import CanvasObject
+
+
+class PairingCode(CanvasObject):
+    def __str__(self):
+        return "{} - {}".format(self.user_id, self.code)