+++ canvasapi/requester.py
@@ -1,5 +1,5 @@
-from datetime import datetime
 import logging
+from datetime import datetime
 from pprint import pformat
 
 import requests
@@ -15,7 +15,6 @@ from canvasapi.exceptions import (
     UnprocessableEntity,
 )
 from canvasapi.util import clean_headers
-
 
 logger = logging.getLogger(__name__)
 
@@ -157,7 +156,7 @@ class Requester(object):
             currently only the POST request of GraphQL is using this parameter.
             For all other methods it's just passed and ignored.
         :type json: `bool`
-        :rtype: str
+        :rtype: :class:`requests.Response`
         """
         full_url = _url if _url else "{}{}".format(self.base_url, endpoint)
 
@@ -218,9 +217,14 @@ class Requester(object):
         )
 
         try:
-            logger.debug("Data: {data}".format(data=pformat(response.json())))
-        except ValueError:
-            logger.debug("Data: {data}".format(data=pformat(response.text)))
+            logger.debug(
+                "Data: {data}".format(data=pformat(response.content.decode("utf-8")))
+            )
+        except UnicodeDecodeError:
+            logger.debug("Data: {data}".format(data=pformat(response.content)))
+        except AttributeError:
+            # response.content is None
+            logger.debug("No data")
 
         # Add response to internal cache
         if len(self._cache) > 4: