+++ tests/test_folder.py
@@ -7,7 +7,7 @@ from canvasapi import Canvas
 from canvasapi.file import File
 from canvasapi.folder import Folder
 from tests import settings
-from tests.util import register_uris, cleanup_file
+from tests.util import cleanup_file, register_uris
 
 
 @requests_mock.Mocker()